# docs/openapi.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "OpenAPI definition",
    "version" : "v0"
  },
  "servers" : [ {
    "url" : "http://localhost",
    "description" : "Generated server url"
  } ],
  "paths" : {
...
    "/api/workspaces/{workspacePublicId}/withdrawals" : {
      "post" : {
        "tags" : [ "withdrawal-controller" ],
        "operationId" : "create_1",
        "parameters" : [ {
          "name" : "workspacePublicId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateWithdrawalRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "description" : "Created",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WithdrawalResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspaces/{workspacePublicId}/withdrawals/{withdrawalPublicId}/release" : {
      "post" : {
        "tags" : [ "withdrawal-controller" ],
        "operationId" : "release",
        "parameters" : [ {
          "name" : "workspacePublicId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "withdrawalPublicId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WithdrawalResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspaces/{workspacePublicId}/withdrawals/{withdrawalPublicId}/mark-seen" : {
      "post" : {
        "tags" : [ "withdrawal-controller" ],
        "operationId" : "markSeen",
        "parameters" : [ {
          "name" : "workspacePublicId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "withdrawalPublicId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WithdrawalResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspaces/{workspacePublicId}/withdrawals/{withdrawalPublicId}/chat/messages" : {
      "post" : {
        "tags" : [ "withdrawal-controller" ],
        "operationId" : "sendChatMessage",
        "parameters" : [ {
          "name" : "workspacePublicId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "withdrawalPublicId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/SendChatMessageRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        }
      }
    },
...
    "/api/workspaces/{workspacePublicId}" : {
      "get" : {
        "tags" : [ "workspace-controller" ],
        "operationId" : "getDetails",
        "parameters" : [ {
          "name" : "workspacePublicId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WorkspaceResponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "workspace-controller" ],
        "operationId" : "delete_1",
        "parameters" : [ {
          "name" : "workspacePublicId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "204" : {
            "description" : "No Content"
          }
        }
      }
    },
    "/api/workspaces/{workspacePublicId}/withdrawals/{withdrawalPublicId}" : {
      "get" : {
        "tags" : [ "withdrawal-controller" ],
        "operationId" : "getDetails_1",
        "parameters" : [ {
          "name" : "workspacePublicId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "withdrawalPublicId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WithdrawalDetailsResponse"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "tags" : [ "withdrawal-controller" ],
        "operationId" : "cancel",
        "parameters" : [ {
          "name" : "workspacePublicId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "withdrawalPublicId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/WithdrawalResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspaces/{workspacePublicId}/withdrawals/{withdrawalPublicId}/receipts/{receiptId}/pdf" : {
      "get" : {
        "tags" : [ "withdrawal-controller" ],
        "operationId" : "getReceiptPdf",
        "parameters" : [ {
          "name" : "workspacePublicId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "withdrawalPublicId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "receiptId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "string",
                  "format" : "byte"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspaces/{workspacePublicId}/withdrawals/completed" : {
      "get" : {
        "tags" : [ "withdrawal-controller" ],
        "operationId" : "getCompleted",
        "parameters" : [ {
          "name" : "workspacePublicId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/WithdrawalResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/workspaces/{workspacePublicId}/withdrawals/active" : {
      "get" : {
        "tags" : [ "withdrawal-controller" ],
        "operationId" : "getActive",
        "parameters" : [ {
          "name" : "workspacePublicId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/WithdrawalResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/workspaces/{workspacePublicId}/foreign-orders/{id}" : {
      "get" : {
        "tags" : [ "foreign-bybit-order-controller" ],
        "operationId" : "getDetails_2",
        "parameters" : [ {
          "name" : "workspacePublicId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        }, {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ForeignBybitOrderResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspaces/{workspacePublicId}/foreign-orders/active" : {
      "get" : {
        "tags" : [ "foreign-bybit-order-controller" ],
        "operationId" : "getActive_1",
        "parameters" : [ {
          "name" : "workspacePublicId",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/ForeignBybitOrderResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "WorkspaceResponse" : {
        "type" : "object",
        "properties" : {
          "publicId" : {
            "type" : "string"
          },
          "name" : {
            "type" : "string"
          },
          "ownerPublicId" : {
            "type" : "string"
          },
          "ownerUsername" : {
            "type" : "string"
          },
          "currentUserRole" : {
            "type" : "string"
          },
          "bybitP2pAdId" : {
            "type" : "string"
          },
          "receiptEmail" : {
            "type" : "string"
          },
          "imapHost" : {
            "type" : "string"
          },
          "imapPort" : {
            "type" : "integer",
            "format" : "int32"
          },
          "imapUsername" : {
            "type" : "string"
          },
          "enabled" : {
            "type" : "boolean"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "CreateWithdrawalRequest" : {
        "type" : "object",
        "properties" : {
          "amountRub" : {
            "type" : "number"
          },
          "recipientPhone" : {
            "type" : "string",
            "minLength" : 1
          },
          "recipientBank" : {
            "type" : "string",
            "minLength" : 1
          },
          "recipientName" : {
            "type" : "string",
            "minLength" : 1
          }
        },
        "required" : [ "amountRub", "recipientBank", "recipientName", "recipientPhone" ]
      },
      "WithdrawalResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "publicId" : {
            "type" : "string"
          },
          "amountRub" : {
            "type" : "number"
          },
          "recipientPhone" : {
            "type" : "string"
          },
          "recipientBank" : {
            "type" : "string"
          },
          "recipientBankTitle" : {
            "type" : "string"
          },
          "recipientName" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "statusTitle" : {
            "type" : "string"
          },
          "attentionRequired" : {
            "type" : "boolean"
          },
          "completionSeen" : {
            "type" : "boolean"
          },
          "queueGroupKey" : {
            "type" : "string"
          },
          "queuePosition" : {
            "type" : "integer",
            "format" : "int32"
          },
          "bybitOrderId" : {
            "type" : "string"
          },
          "bybitOrderAmountRub" : {
            "type" : "number"
          },
          "bybitOrderQuantityUsdt" : {
            "type" : "number"
          },
          "bybitOrderFeeUsdt" : {
            "type" : "number"
          },
          "bybitOrderTotalUsdt" : {
            "type" : "number"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "queuedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "publishedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "orderFoundAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "requisitesSentAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "paidAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "verificationStartedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "completedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "cancelledAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "lastError" : {
            "type" : "string"
          },
          "lastWarning" : {
            "type" : "string"
          },
          "createdByUsername" : {
            "type" : "string"
          },
          "canCancel" : {
            "type" : "boolean"
          },
          "canRelease" : {
            "type" : "boolean"
          }
        }
      },
      "SendChatMessageRequest" : {
        "type" : "object",
        "properties" : {
          "message" : {
            "type" : "string",
            "maxLength" : 1000,
            "minLength" : 0
          }
        },
        "required" : [ "message" ]
      },
...
      "ChatMessageLogResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string"
          },
          "bybitOrderId" : {
            "type" : "string"
          },
          "messageIndex" : {
            "type" : "integer",
            "format" : "int32"
          },
          "messageText" : {
            "type" : "string"
          },
          "direction" : {
            "type" : "string"
          },
          "authorName" : {
            "type" : "string"
          },
          "contentType" : {
            "type" : "string"
          },
          "status" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "error" : {
            "type" : "string"
          }
        }
      },
      "EmailReceiptCheckResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "bybitOrderId" : {
            "type" : "string"
          },
          "emailMessageId" : {
            "type" : "string"
          },
          "emailFrom" : {
            "type" : "string"
          },
          "emailSubject" : {
            "type" : "string"
          },
          "emailReceivedAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "pdfFilename" : {
            "type" : "string"
          },
          "pdfAvailable" : {
            "type" : "boolean"
          },
          "parsedStatus" : {
            "type" : "string"
          },
          "parsedAmountRub" : {
            "type" : "number"
          },
          "parsedRecipientPhone" : {
            "type" : "string"
          },
          "parsedRecipientBank" : {
            "type" : "string"
          },
          "parsedRecipientName" : {
            "type" : "string"
          },
          "parsedOperationDate" : {
            "type" : "string"
          },
          "parsedOperationId" : {
            "type" : "string"
          },
          "parsedReceiptNumber" : {
            "type" : "string"
          },
          "verificationStatus" : {
            "type" : "string"
          },
          "verificationError" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "WithdrawalDetailsResponse" : {
        "type" : "object",
        "properties" : {
          "withdrawal" : {
            "$ref" : "#/components/schemas/WithdrawalResponse"
          },
          "events" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/WithdrawalEventResponse"
            }
          },
          "chatMessages" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ChatMessageLogResponse"
            }
          },
          "receiptChecks" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/EmailReceiptCheckResponse"
            }
          }
        }
      },
      "WithdrawalEventResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "eventType" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "payloadJson" : {
            "type" : "string"
          },
          "actorType" : {
            "type" : "string"
          },
          "actorUsername" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "ForeignBybitOrderResponse" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "integer",
            "format" : "int64"
          },
          "bybitOrderId" : {
            "type" : "string"
          },
          "amountRub" : {
            "type" : "number"
          },
          "bybitStatus" : {
            "type" : "string"
          },
          "reason" : {
            "type" : "string"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "updatedAt" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
...
    }
  }
}